Resolve "saving data to databse"

Closes #28

See merge request !29

+++ app/src/main/java/com/example/olive/carbon_tracker/Model/Vehicle.java
@@ -9,15 +9,15 @@ public class Vehicle {
     private String Model;
     private int Year;
     private int iconId = R.drawable.old_vintage_car;
-    private int Highway08;
-    private int City08;
+    private double Highway08;
+    private double City08;
 
     private String FuelType;
 
     private int cityDistance = 0;
     private int hwyDistance = 0;
 
-    public Vehicle(String name, String make,String model, int year,int city08, int highway08, String fuelType ){
+    public Vehicle(String name, String make,String model, int year,double city08, double highway08, String fuelType ){
         this.Name = name;
         this.Make = make;
         this.Model = model;
@@ -28,11 +28,11 @@ public class Vehicle {
 
     }
 
-    public int getHighway08() {
+    public double getHighway08() {
         return Highway08;
     }
 
-    public int getCity08() {
+    public double getCity08() {
         return City08;
     }
